refactor: Remove top navigation bar and integrate data collection features
- Remove Navigation component and its usage from layout
- Integrate advanced data collection functionality into existing data-upload page
- Add drag & drop file upload, Excel processing, and system status display
- Maintain CommonShell wrapper for consistent UI structure
- All build tests passing (frontend build, lint, type-check)
- Remove standalone data-collection page (functionality merged)
