+++ frontend/public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const c=(c,a)=>(c=new URL(c+".js",a).href,s[c]||new Promise(s=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=s,document.head.appendChild(e)}else e=c,importScripts(c),s()}).then(()=>{let e=s[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e}));self.define=(a,i)=>{const t=e||("document"in self?document.currentScript.src:"")||location.href;if(s[t])return;let n={};const r=e=>c(e,t),o={module:{uri:t},exports:n,require:r};s[t]=Promise.all(a.map(e=>o[e]||r(e))).then(e=>(i(...e),n))}}define(["./workbox-977fb9bf"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"cdbe1ab85db8b013d11ee68c8c26c49a"},{url:"/_next/static/chunks/124-f5b9957d8080f4f3.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/147b5626-eda80bf06e788917.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/590-038d87e4285f47f5.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/691-4f3f394222c29ae6.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/757-71f279ffbaef063d.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/ad1ff839-92e83b5aab47e8bf.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/(protected)/cbam/page-8c294ad415d9cf61.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/(protected)/dashboard/page-d66441d8d98331a8.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/(protected)/data-upload/page-2764ebd3b57e61cd.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/(protected)/lca/page-57b6c55c812b80cf.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/(protected)/settings/page-32cb3967d077e5b0.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/(public)/landing/page-1d55ba8fc74d651d.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/(public)/register/page-9209bcb25c5995b3.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/_not-found/page-958dc03f549bafe5.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/data-collection/page-5adaf9cf42c9849c.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/debug/page-43bc24ed0337993c.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/layout-25a0657ade00476e.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/app/page-cc60ead3af32c60c.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/framework-6e06c675866dc992.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/main-app-205ad2eeb3078426.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/main-bbcacd86e8fc73d8.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/pages/_app-1e0813bbe740beb7.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/pages/_error-7730e932b281c983.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-4da7738d6019a505.js",revision:"mV7kjpCXzGQozRyZgNZ4T"},{url:"/_next/static/css/35b48a343af3078f.css",revision:"35b48a343af3078f"},{url:"/_next/static/mV7kjpCXzGQozRyZgNZ4T/_buildManifest.js",revision:"97e22fcc3f8be8f35ac7d89770b9078d"},{url:"/_next/static/mV7kjpCXzGQozRyZgNZ4T/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/media/26a46d62cd723877-s.woff2",revision:"befd9c0fdfa3d8a645d5f95717ed6420"},{url:"/_next/static/media/55c55f0601d81cf3-s.woff2",revision:"43828e14271c77b87e3ed582dbff9f74"},{url:"/_next/static/media/581909926a08bbc8-s.woff2",revision:"f0b86e7c24f455280b8df606b89af891"},{url:"/_next/static/media/8e9860b6e62d6359-s.woff2",revision:"01ba6c2a184b8cba08b0d57167664d75"},{url:"/_next/static/media/97e0cb1ae144a2a9-s.woff2",revision:"e360c61c5bd8d90639fd4503c829c2dc"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/_next/static/media/e4af272ccee01ff0-s.p.woff2",revision:"65850a373e258f1c897a2b3d75eb74de"},{url:"/apple-touch-icon.svg",revision:"4b61b033bb139fab47e6181911182b43"},{url:"/favicon.svg",revision:"a83a66200ef44427c23051773c7a0d60"},{url:"/icon-192x192.svg",revision:"deb0a3fae595a0564b83759f7e7ff5d7"},{url:"/icon-512x512.svg",revision:"36dcaca1f2d2e35efe893bd4373b6c94"},{url:"/icon.svg",revision:"36dcaca1f2d2e35efe893bd4373b6c94"},{url:"/manifest.json",revision:"002c41d4543ba55ea03458cdf458583f"},{url:"/offline.html",revision:"afd584b2df9d6c316bfb16a340109341"},{url:"/robots.txt",revision:"119243b37e2db160c999906d9cc98c94"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:c,state:a})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/api\.greensteel\.site/,new e.NetworkFirst({cacheName:"api-cache",plugins:[new e.ExpirationPlugin({maxEntries:100,maxAgeSeconds:86400}),new e.CacheableResponsePlugin({statuses:[0,200]})]}),"GET"),e.registerRoute(/\.(?:png|jpg|jpeg|svg|gif|webp|avif)$/,new e.CacheFirst({cacheName:"image-cache",plugins:[new e.ExpirationPlugin({maxEntries:200,maxAgeSeconds:2592e3})]}),"GET"),e.registerRoute(/\.(?:js|css)$/,new e.StaleWhileRevalidate({cacheName:"static-resources",plugins:[new e.ExpirationPlugin({maxEntries:100,maxAgeSeconds:604800})]}),"GET")});
+if(!self.define){let e,s={};const n=(n,i)=>(n=new URL(n+".js",i).href,s[n]||new Promise(s=>{if("document"in self){const e=document.createElement("script");e.src=n,e.onload=s,document.head.appendChild(e)}else e=n,importScripts(n),s()}).then(()=>{let e=s[n];if(!e)throw new Error(`Module ${n} didn’t register its module`);return e}));self.define=(i,a)=>{const c=e||("document"in self?document.currentScript.src:"")||location.href;if(s[c])return;let t={};const r=e=>n(e,c),d={module:{uri:c},exports:t,require:r};s[c]=Promise.all(i.map(e=>d[e]||r(e))).then(e=>(a(...e),t))}}define(["./workbox-977fb9bf"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"a3e6251a769b89c470ab97772d8df790"},{url:"/_next/static/chunks/147b5626-eda80bf06e788917.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/590-038d87e4285f47f5.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/691-4f3f394222c29ae6.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/954-011f997324349b75.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/ad1ff839-92e83b5aab47e8bf.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/app/(protected)/cbam/page-8c294ad415d9cf61.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/app/(protected)/dashboard/page-8559fac1fb4a168f.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/app/(protected)/data-upload/page-8d8374b8e59fa42e.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/app/(protected)/lca/page-57b6c55c812b80cf.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/app/(protected)/settings/page-32cb3967d077e5b0.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/app/(public)/landing/page-4de33ede389facf4.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/app/(public)/register/page-9209bcb25c5995b3.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/app/_not-found/page-958dc03f549bafe5.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/app/debug/page-43bc24ed0337993c.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/app/layout-e1009d618c8660e0.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/app/page-ceb31d530ab29f94.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/framework-6e06c675866dc992.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/main-6519ebbec0a1c2d5.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/main-app-205ad2eeb3078426.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/pages/_app-1e0813bbe740beb7.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/pages/_error-7730e932b281c983.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-4da7738d6019a505.js",revision:"rICHhOGmDGTsy9z_NnGld"},{url:"/_next/static/css/83b48e48a318570c.css",revision:"83b48e48a318570c"},{url:"/_next/static/media/26a46d62cd723877-s.woff2",revision:"befd9c0fdfa3d8a645d5f95717ed6420"},{url:"/_next/static/media/55c55f0601d81cf3-s.woff2",revision:"43828e14271c77b87e3ed582dbff9f74"},{url:"/_next/static/media/581909926a08bbc8-s.woff2",revision:"f0b86e7c24f455280b8df606b89af891"},{url:"/_next/static/media/8e9860b6e62d6359-s.woff2",revision:"01ba6c2a184b8cba08b0d57167664d75"},{url:"/_next/static/media/97e0cb1ae144a2a9-s.woff2",revision:"e360c61c5bd8d90639fd4503c829c2dc"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/_next/static/media/e4af272ccee01ff0-s.p.woff2",revision:"65850a373e258f1c897a2b3d75eb74de"},{url:"/_next/static/rICHhOGmDGTsy9z_NnGld/_buildManifest.js",revision:"97e22fcc3f8be8f35ac7d89770b9078d"},{url:"/_next/static/rICHhOGmDGTsy9z_NnGld/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/apple-touch-icon.svg",revision:"4b61b033bb139fab47e6181911182b43"},{url:"/favicon.svg",revision:"a83a66200ef44427c23051773c7a0d60"},{url:"/icon-192x192.svg",revision:"deb0a3fae595a0564b83759f7e7ff5d7"},{url:"/icon-512x512.svg",revision:"36dcaca1f2d2e35efe893bd4373b6c94"},{url:"/icon.svg",revision:"36dcaca1f2d2e35efe893bd4373b6c94"},{url:"/manifest.json",revision:"002c41d4543ba55ea03458cdf458583f"},{url:"/offline.html",revision:"afd584b2df9d6c316bfb16a340109341"},{url:"/robots.txt",revision:"119243b37e2db160c999906d9cc98c94"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:n,state:i})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/api\.greensteel\.site/,new e.NetworkFirst({cacheName:"api-cache",plugins:[new e.ExpirationPlugin({maxEntries:100,maxAgeSeconds:86400}),new e.CacheableResponsePlugin({statuses:[0,200]})]}),"GET"),e.registerRoute(/\.(?:png|jpg|jpeg|svg|gif|webp|avif)$/,new e.CacheFirst({cacheName:"image-cache",plugins:[new e.ExpirationPlugin({maxEntries:200,maxAgeSeconds:2592e3})]}),"GET"),e.registerRoute(/\.(?:js|css)$/,new e.StaleWhileRevalidate({cacheName:"static-resources",plugins:[new e.ExpirationPlugin({maxEntries:100,maxAgeSeconds:604800})]}),"GET")});
